+++ common/src/main/java/com/binaris/wizardry/client/EBClientEventHelper.java
@@ -1,7 +1,9 @@
 package com.binaris.wizardry.client;
 
 import com.binaris.wizardry.api.content.event.EBClientTickEvent;
+import com.binaris.wizardry.api.content.event.EBEntityJoinLevelEvent;
 import com.binaris.wizardry.api.content.event.EBLivingTick;
+import com.binaris.wizardry.client.effect.ArcaneLockRender;
 import com.binaris.wizardry.client.sound.SoundLoop;
 import com.binaris.wizardry.core.event.WizardryEventBus;
 import com.binaris.wizardry.setup.registries.client.EBKeyBinding;
@@ -14,6 +16,7 @@ public final class EBClientEventHelper {
         WizardryEventBus bus = WizardryEventBus.getInstance();
         onLivingTickEvent(bus);
         onClientTick(bus);
+        onJoin(bus);
     }
 
     private static void onLivingTickEvent(WizardryEventBus bus) {
@@ -24,6 +27,10 @@ public final class EBClientEventHelper {
         bus.register(EBClientTickEvent.class, SoundLoop::onClientTick);
         bus.register(EBClientTickEvent.class, EBKeyBinding::onClientTick);
         bus.register(EBClientTickEvent.class, ScreenShakeHandler::onClientTick);
+    }
+
+    private static void onJoin(WizardryEventBus bus) {
+        bus.register(EBEntityJoinLevelEvent.class, ArcaneLockRender::onJoin);
     }
 
 }